sketch sample parsing

+++ src/read.h
@@ -25,7 +25,7 @@ Int24 read_i24(const std::vector<Byte>& bytes,
                size_t& index,
                IndexPolicy policy);
 
-i32 read_sample(u8 bit_depth,
+i64 read_sample(u8 bit_depth,
                 const std::vector<Byte>& bytes,
                 size_t& index,
                 IndexPolicy policy);
